test(mcp): move directory command doctor regression to a focused file
